feat: add event carousel slider with event detail page
