+++ lib/screens/login_screen/widgets/skip_and_next_row.dart
@@ -26,7 +26,7 @@ class SkipAndNextRow extends StatelessWidget {
             child: ElevatedButton(
               onPressed: () {
                 Navigator.of(context).push(
-                  MaterialPageRoute(builder: (context) =>  MainScreen()),
+                  MaterialPageRoute(builder: (context) => const MainScreen()),
                 );
               },
               child: IconButton(
